v0.15.3 - Correctif comparaison stricte float/int (calendrier RDV toujours vide)

+++ includes/rest/class-grc-rest-rdv.php
@@ -89,7 +89,7 @@ class GRC_REST_RDV {
 
 		$results = [];
 		foreach ( $rows as $c ) {
-			$duree_minutes = round( ( strtotime( $c->fin ) - strtotime( $c->debut ) ) / 60 );
+			$duree_minutes = (int) round( ( strtotime( $c->fin ) - strtotime( $c->debut ) ) / 60 );
 			if ( $duree_filtre && $duree_minutes !== $duree_filtre ) {
 				continue;
 			}
